feat: we now write dependencies to package.json
We update on changes, theoretically. This required adding "initial
settings" (everything off because of first run) in addition to
"default settings" (the things we'll turn on for you on first run).

Also, methodically fixed every problem in the way of making
`npm i; node boltzmann.js` work in the scaffolded app. It now
responds to hello/world and to ping & status!

There are still stray <? ?> in the template file that will need
to be cleaned up, however, so it won't fully work yet.