doc: Rework the middleware concept document.
Link off to nearly-empty Context, Response, and middleware reference
documentation. Use TypeScript type definitions to give us names for the various
components of a given middleware. Use an onion analogy, because I can't be
helped. An illustration of this would be useful.

This also adds a "section.html" template to the after-dark, which
is just a copy of "index.html".