feat: csrf protection middlware
This implementation uses the double-submit pattern to dodge around having
to store state in a user session, since we don't have a built-in session store
yet. It uses the same approach as the connect-style `csurf` middleware, and
bases its implementation on the same underlying random token implementation.
We store a secret in a signed cookie for the user, and verify that a) the cookie
signature is good, and that b) the submitted token is valid for that secret.

Configuration options:
cookieSecret: the secret used to sign the cookie
param: a body param to look for the csrf token in
header: a response header to look for the csrf token in; is consulted first if both are used

Wrote tests for it that verify we ignore read http verbs, we enforce it
on write verbs, we look in the body param specified, we look in the header
specified, and we bounce on bad tokens & bad cookie signatures.

Added this as a command-line option. Our current approach to feature options
is now officially unwieldy and the next one will trigger me to rewrite it.

Tweaked `--selftest` so it turns on all options, and cleaned up the
self test. Also, made the test script more reliable in my testing by
using pkill to detect boltzmann processes.