feat: improved docs, templates, and examples for websites
New projects using templates get a template starter!
If you scaffold out anything with templates turned on, we
give you an example index.html template file and a starter route
that uses it to greet the world.

Wrote more concepts documentation for websites.  Documented cookies,
csrf protection, templates, and the debug error page.

Expanded on the middleware template we render for you with more
examples of how to configure the middleware you've asked to have installed.

Added a shortcode for resizing images in docs files.

Linked the readme to our latest docs.

Added a very simple website with two routes to the examples directory.
It's mostly notable for the comments in the files.