Add ESM support for Node 14.
ESM is a special flag. It is not implied by "--all" or "--selftest"
and is tested separately.

Turning the flag on puts the application in "module mode" by setting
the "type" flag in `package.json`. `boltzmann.js` then expects to load
all application files (middleware, handlers, etc.) via `import()`.

The upside: it works! Import syntax makes handlers definition much nicer
by baking the export status into the definition of the handler function.
This also puts all files into strict mode by default, which is good for
usabiliy and performance.

The downside is that it forks exports and imports in all of our generated
files. I think, medium-to-long-term, it's worthwhile: ESM is the future of JS,
t and Node 14 forces the issue. Ultimately, we should default to ESM. HOWEVER,
this complicates hot-reloading slightly: it means we end up having to build
it twice (or have one mode that supports it, and one that doesn't.)