Docs on the website features (#31)
* feat: improved docs, templates, and examples for websites

New projects using templates get a template starter!
If you scaffold out anything with templates turned on, we
give you an example index.html template file and a starter route
that uses it to greet the world.

Wrote more concepts documentation for websites.  Documented cookies,
csrf protection, templates, and the debug error page.

Expanded on the middleware template we render for you with more
examples of how to configure the middleware you've asked to have installed.

Added a shortcode for resizing images in docs files.

Linked the readme to our latest docs.

Added a very simple website with two routes to the examples directory.
It's mostly notable for the comments in the files.

* chore: clean up the now-unused after-dark theme

* Editing pass on the website docs.

* Drag the concepts docs folder closer to the goal IA.

* Test cleanup.

Fixed lint errors with the output of the scaffolder in
the website example. Added an almost useless test for
the website example.

Perhaps controversially, turned off the githubci feature
for all example dir test runs.

* Don't blow away an index.html template if it exists.

* We don't say "mount" any more; we say "attach".

That word goes into the dustbin of history next to "decorator".

Expanded on the concept of "dev mode" a bit.

* Made the example a little more initial-use friendly.

Did the same to the template. Also, added an example of
using bole to log.

* /hello/world considered better most of the time

except in the example