Some documentation work on the cli options & the landing page (#33)
Improved the CLI option docs.

Scratched an itch I've had for a while by making the docs landing page not a rehash of the readme. It's no longer a rehash of the repo readme because a) the repo readme is shorter (it's the manifesto) and the landing page goes into more detail and does some linking to useful doc starting points.

Bonus: the default scaffolded project almost passes lint. There's one bug in that the middleware template pulls in boltzmann, but the default option set doesn't enable anything that uses the variable.

Alphabetized the displayed features list.