Fix issue #39: bad middleware config should fail fast (#41)
* Fix issue #39: bad middleware config should fail fast

applyXFO() now fails fast. Added a test. applyXFO() is now scaffolded only if templates are enabled.

template() now fails informatively if `paths` cannot be an array of strings instead of failing
uniformatively. Added a test.

authenticateJWT() now wants an array of strings for its algorithms config. It doesn't attempt to
validate the algorithm choices. It also requires that you set something in AUTHENTICATION_KEY so it
can fail fast if you've forgotten to. This required some test updates.

Extended the scaffold to include more middleware config examples. The scaffolded project runs out of
the box, though it does not yet quite pass lint out of the box. (Apparently my favorite misspelling
this week is "middlware".)

Bumped versions of pg & find-my-way to the latest. The version bump of pg comes with some risk
because of bugs, but is required for use to work on node 14.