fix: throw out bad sessions instead of crashing (#52)
* fix: throw out bad sessions instead of crashing

We now log and toss out sessions that fail to decrypt with the
current secret or are malformed after decrypting.

Ideally we'll implement secret rotations and fallbacks. In the
short term, however, I think we should handle this more gracefully.

* Better handling for malformed sessions.

We now throw BadSessionError and 400 if the contents of the
sealed cookie were bad. Corrected the misspelling of `BadSessionError`,
which had more Rs than any error class name should have.