feat(oauth): Refactor oauth config & document it
This work is untested and incomplete, but indicates the
direction I'd like to take the oauth config. We now massage
and validate configuration in the central oauth() middleware
before attaching the three sub-middlewares that do all the work.
This has the salutary effect of drying up some repeated config
arg processing and some throwing-on-missing-config.

Started documenting the config, but interrupted that work to do
the refactor once I understood what was going on.