feat: tabular output for dependencies & features
Reworked the NPM dependency output to be columnar and more terse.  This output
is now suppressed on first run of the scaffold, but shown on subsequent runs to
highlight changes.  If verbosity is at debug level or higher, we always show
the dependency list.

Gave the features list the same treatment, though it is always displayed at the
end of a run. Both types of output use prettytables for display, which allowed
me to put the gridding logic into a re-used function.

We can now tweak this until we're happy.

Added a feature-list getter to settings, and made the display formatter use
that.

Added --quiet as a synonym for --silent.

Removed now-unused deps.

Upgraded tera to version 1.6.0.