feat: run-script messaging is now more compact
To fix a problem with confusing messaging for the `mime' dependency,
implemented an `any_of` clause for the When struct. This enables a dependency
if any of its possible prerequisites are present. Put it to work with the
`mime` dep, which is pulled in when any of esbuild or staticfiles is enabled.

This made the wants-a-dep repeated check complex enough that it warranted
moving to a function, so this is now in `dep_wanted()`.

And then got to the thing I really wanted to do, which was make the run script
management output more compact. We note changes, additions, and when we want to
manage something but are declining because the user overrode us.

One last change: First-run output now gives a total dependencies-added count,
if logging is at info level. It continues to give the full output if the log
level is chattier.