feat: run-script messaging is now more compact (#76)
* fix: we were too aggro about stripping whitespace

The initial - in {%- endif -%} joined lines together, and boy was
the typescript transpiler unhappy about that.

* feat: run-script messaging is now more compact

To fix a problem with confusing messaging for the `mime' dependency,
implemented an `any_of` clause for the When struct. This enables a dependency
if any of its possible prerequisites are present. Put it to work with the
`mime` dep, which is pulled in when any of esbuild or staticfiles is enabled.

This made the wants-a-dep repeated check complex enough that it warranted
moving to a function, so this is now in `dep_wanted()`.

And then got to the thing I really wanted to do, which was make the run script
management output more compact. We note changes, additions, and when we want to
manage something but are declining because the user overrode us.

One last change: First-run output now gives a total dependencies-added count,
if logging is at info level. It continues to give the full output if the log
level is chattier.

* Fix: make nodemon watch ts files for typescript apps

* add set -e

* (cargo-release) version 0.3.0-rc.1

* `wants_dep()` is now a function on When

and is now spelled `are_satisfied_by()` because the variables seem
to be all plurals. E.g., `preconditions.are_satisfied_by(settings)`.

Made the file renderer use this test instead of its own. This somewhat
damages the usefulness of trace-level output, but since I am probably
the only consumer of that output, I can deal.

Fix more whitespace collapsing imperfections.

* Javascript is the nicest, cleanest, clearest, most wonderful programming language I've ever known in my life.

Co-authored-by: Chris Dickinson <christopher.s.dickinson@gmail.com>