feat: documentation & an example for TypeScript (#78)
* feat: documentation & an example for TypeScript

There is now documentation for the typescript cli option.

Cleaned up the tsconfig a bit to remove statements that were redundant
with the base node14 config.

TypeScript & esm options are now marked as mutually exclusive.  It is technically
possible to use them together, by using the experimental module loader like this:
`node --loader ts-node/esm -r dotenv/config` but the routes --eval trick doesn't
work. Punting on interop here for now.

The typescript post-install hook now produces working projects that can be executed with
`node target/release/boltzmann.js`. There is one thing I think is less than great,
which is that boltzmann defaults to reading the package name from package.json,
which means that it needs to be copied into the built project even though it won't work.
There are yaks yet to shave here.

Added a `--quiet` alias for `--silent` because half the time I type one and the other
half of the time the other.

Updated cli usage examples in the docs. Mentioned the typescript feature in the top-level
README.

* Fix a copy-paste error in a run script.