fix #77: B=1 considered baffling to Windows (#79)
fix #77: B=1 considered baffling to Windows

We now track run scripts through versions of Boltzmann. If a standard run
script target (one not prefixed with `boltzmann:`) if already set, we compare
it to our past history of values for that run script to decide if we're
managing it.

`runscripts.ron` now records my best understanding of the history of our run
scripts. We've only really changed two of them over time, as far as I can tell.

Updated all examples with fresh run scripts, cleaning up as needed. In
particular, B=1 is removed from all of them. Once this was gone, tests on
Windows started failing because file globbing doesn't work as expected. Cleaned
that up.

Took this opportunity to improve output a little bit. We now tell you if you're
upgrading an installation by comparing the previous version with the one you're
applying.