Fix broken tests, improve test output.
Tests broke at some point due to dependencies upgrading; in particular
TAP required an explicit version bump. Once I did this I realized another
test was failing in bin/test, but it was difficult to track down. I added
wrapping output for tests to determine which section of tests failed, and
bumped the verbosity of the prod_only_build_ts test to 3 so I could debug
why it was failing. It turns out that we had a type regression due to
dependency updates. This commit contains a fix (see oauth.ts).

Finally, I removed the jarring output that results from `kill %1` in our
"startup" tests: I manually add a "shut down the server" endpoint to handlers
and call it when we're done with the server. Time will tell if that's a good
idea or not!

Finally, I bumped TAP in dependencies.ron and it seems like we're passing tests
now.